--- tests/expected/function-contract/interior-mutability/transmute/cell_stub.expected
@@ -1,13 +1,0 @@
-assertion\
-- Status: SUCCESS\
-- Description: "|_| old(*unsafe{ im.x.expose() } + *unsafe{ im.x.expose() }) ==\
-*unsafe{ im.x.expose() }"\
-
-assertion\
-- Status: SUCCESS\
-- Description: "|_|
-old(*unsafe{ im.x.expose() } + *unsafe{ im.x.expose() } +\
-*unsafe{ im.x.expose() } + *unsafe{ im.x.expose() }) ==\
-*unsafe{ im.x.expose() }"\
-
-VERIFICATION:- SUCCESSFUL
