+++ tests/expected/function-contract/interior-mutability/whole-struct/cell.expected
@@ -1,0 +1,6 @@
+assertion\
+- Status: SUCCESS\
+- Description: "|_| im.x.get() < 101"\
+in function modify
+
+VERIFICATION:- SUCCESSFUL
