+++ tests/expected/function-contract/interior-mutability/whole-struct/oncecell.expected
@@ -1,0 +1,6 @@
+assertion\
+- Status: SUCCESS\
+- Description: "|_| im.x.get().is_some()"\
+in function modify
+
+VERIFICATION:- SUCCESSFUL
