+++ tests/expected/function-contract/interior-mutability/whole-struct/refcell.expected
@@ -1,0 +1,6 @@
+assertion\
+- Status: SUCCESS\
+- Description: "|_| unsafe{ *im.x.as_ptr() } < 101"\
+in function modify
+
+VERIFICATION:- SUCCESSFUL
