+++ tests/expected/function-contract/interior-mutability/whole-struct/refcell.rs
@@ -2,26 +2,26 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 // kani-flags: -Zfunction-contracts
 
-/// Mutating Cell via as_ptr in contracts
-use std::cell::Cell;
+/// The objective of this test is to check the modification of a RefCell used as interior mutability in an immutable struct
+
+use std::cell::RefCell;
 
 /// This struct is contains Cell which can be mutated
 struct InteriorMutability {
-    x: Cell<u32>,
+    x: RefCell<u32>,
 }
 
-/// contracts need to access im.x internal data through the api im.x.as_ptr
-#[kani::requires(im.x.get() < 100)]
-#[kani::modifies(im.x.as_ptr())]
-#[kani::ensures(|_| im.x.get() < 101)]
+#[kani::requires(unsafe{*im.x.as_ptr()} < 100)]
+#[kani::modifies(&im.x)]
+#[kani::ensures(|_| unsafe{*im.x.as_ptr()} < 101)]
 ///im is an immutable reference with interior mutability
 fn modify(im: &InteriorMutability) {
     // valid rust methodology for getting and setting value without breaking encapsulation
-    im.x.set(im.x.get() + 1)
+    im.x.replace_with(|&mut old| old + 1);
 }
 
 #[kani::proof_for_contract(modify)]
 fn harness_for_modify() {
-    let im: InteriorMutability = InteriorMutability { x: Cell::new(kani::any()) };
+    let im: InteriorMutability = InteriorMutability { x: RefCell::new(kani::any()) };
     modify(&im)
 }
