+++ tests/expected/function-contract/interior-mutability/whole-struct/unsafecell.expected
@@ -1,0 +1,6 @@
+assertion\
+- Status: SUCCESS\
+- Description: "|_| unsafe{ *im.x.get() } < 101"\
+in function modify
+
+VERIFICATION:- SUCCESSFUL
